+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ This project adheres to `Semantic Versioning <http://semver.org/>`_.
 * Added ``--max-headers`` to allow setting the max header limit.
 * Added ``--compress``.
 * Added ``https`` alias command with ``https://`` as the default scheme.
+* Fixed an exception when ``stdin`` was a closed fd.
 
 
 `1.0.3`_ (2019-08-26)